Safari-609.1.11.1/LayoutTests/canvas/philip/tests/2d.imageData.get.order.alpha.html

<!DOCTYPE html>
<title>Canvas test: 2d.imageData.get.order.alpha</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = 'rgba(0, 0, 0, 0.5)';
ctx.fillRect(0, 0, 100, 50);
var imgdata = ctx.getImageData(0, 0, 10, 10);
_assert(imgdata.data[3] < 200, "imgdata.data[\""+(3)+"\"] < 200");
_assert(imgdata.data[3] > 100, "imgdata.data[\""+(3)+"\"] > 100");


});
</script>
