LayoutTests/canvas/philip/tests/2d.imageData.get.order.alpha-expected.txt

< [index] >
2d.imageData.get.order.alpha
getImageData() returns A in the fourth component
References: 2d.pixelarray.order
Actual output:
Passed
